tests(web): update unit tests after refactoring
